arranged directories
