+++ tests/CMakeLists.txt
@@ -45,10 +45,10 @@ add_executable(Prim graphs/mst/prim.cc)
 target_link_libraries(Prim gtest_main)
 gtest_discover_tests(Prim)
 
-add_executable(ZAlgo ./z.cc)
+add_executable(ZAlgo strings/z.cc)
 target_link_libraries(ZAlgo gtest_main)
 gtest_discover_tests(ZAlgo)
 
-add_executable(PrimeSieve ./prime_sieve.cc)
+add_executable(PrimeSieve math/numtheory/prime_sieve.cc)
 target_link_libraries(PrimeSieve gtest_main)
 gtest_discover_tests(PrimeSieve)
