+++ tests/strings/z.cc
